+++ apps/client/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@react-oauth/google": "0.12.1",
     "@tanstack/react-query": "5.62.11",
+    "@waffle/api": "workspace:*",
     "@waffle/design-system": "workspace:*",
     "react": "18.3.1",
     "react-dom": "18.3.1",
