+++ apps/client/src/service/authService.ts
@@ -1,30 +1,46 @@
+import type { Apis } from '@waffle/api';
+
 import type { ServiceResponse } from '@/entities/response';
 import type { User } from '@/entities/user';
-import type { Apis } from '@/shared/api';
 import type { TokenLocalStorage } from '@/shared/token/localstorage';
 import type { TokenState } from '@/shared/token/state';
 
 export type AuthService = {
-  localSignUp({
-    username,
-    localId,
-    password,
-    snuMail,
+  signUp({
+    authType,
+    info,
   }: {
-    username: string;
-    snuMail: string;
-    localId: string;
-    password: string;
+    authType: 'LOCAL_APPLICANT' | 'SOCIAL_APPLICANT';
+    info:
+      | {
+          name: string;
+          localLoginId: string;
+          snuMail: string;
+          password: string;
+        }
+      | {
+          provider: 'google';
+          snuMail: string;
+          token: string;
+        };
   }): ServiceResponse<{
     user: Pick<User, 'id' | 'username' | 'isAdmin'>;
     accessToken: string;
   }>;
-  localSignIn({
-    localId,
-    password,
+  signIn({
+    authType,
+    info,
   }: {
-    localId: string;
-    password: string;
+    authType: 'LOCAL' | 'SOCIAL';
+    info:
+      | {
+          localLoginId: string;
+          password: string;
+        }
+      | {
+          provider: 'google';
+          token: string;
+        };
   }): ServiceResponse<{
     user: Pick<User, 'id' | 'username' | 'isAdmin'>;
     accessToken: string;
@@ -34,24 +50,6 @@ export type AuthService = {
   }: {
     token: string;
   }): ServiceResponse<{ googleEmail: string }>;
-  googleSignUp({
-    snuMail,
-    googleAccessToken,
-  }: {
-    snuMail: string;
-    googleAccessToken: string;
-  }): ServiceResponse<{
-    user: Pick<User, 'id' | 'username' | 'isAdmin'>;
-    accessToken: string;
-  }>;
-  googleSignIn({
-    googleAccessToken,
-  }: {
-    googleAccessToken: string;
-  }): ServiceResponse<{
-    user: Pick<User, 'id' | 'username' | 'isAdmin'>;
-    accessToken: string;
-  }>;
   sendEmailCode({ snuMail }: { snuMail: string }): ServiceResponse<void>;
   verifyCode({
     snuMail,
@@ -67,30 +65,6 @@ export type AuthService = {
   }): ServiceResponse<void>;
   reissueAccessToken(): ServiceResponse<{ accessToken: string }>;
   logout({ token }: { token: string }): ServiceResponse<void>;
-  addLocalSignUp({
-    username,
-    localId,
-    password,
-    snuMail,
-  }: {
-    username: string;
-    snuMail: string;
-    localId: string;
-    password: string;
-  }): ServiceResponse<{
-    user: Pick<User, 'id' | 'username' | 'isAdmin'>;
-    accessToken: string;
-  }>;
-  addGoogleSignUp({
-    snuMail,
-    googleAccessToken,
-  }: {
-    snuMail: string;
-    googleAccessToken: string;
-  }): ServiceResponse<{
-    user: Pick<User, 'id' | 'username' | 'isAdmin'>;
-    accessToken: string;
-  }>;
 };
 
 export const implAuthService = ({
@@ -102,9 +76,9 @@ export const implAuthService = ({
   tokenLocalStorage: TokenLocalStorage;
   tokenState: TokenState;
 }): AuthService => ({
-  localSignUp: async ({ username, localId, password, snuMail }) => {
-    const body = { username, localId, password, snuMail };
-    const { status, data } = await apis['POST /user/signup/local']({ body });
+  signUp: async ({ authType, info }) => {
+    const body = { authType, info };
+    const { status, data } = await apis['POST /user/signup']({ body });
 
     if (status === 200) {
       const token = data.accessToken;
@@ -119,49 +93,15 @@ export const implAuthService = ({
     }
     return { type: 'error', status, message: data.error };
   },
-  localSignIn: async ({ localId, password }) => {
-    const body = { localId, password };
-    const { status, data } = await apis['POST /user/signin/local']({ body });
+  signIn: async ({ authType, info }) => {
+    const body = { authType, info };
+    const { status, data } = await apis['POST /user/signin']({ body });
 
     if (status === 200) {
       const token = data.accessToken;
 
       tokenLocalStorage.setToken({ token });
       tokenState.setToken({ token });
-
-      return {
-        type: 'success',
-        data,
-      };
-    }
-    return { type: 'error', status, message: data.error };
-  },
-  googleSignUp: async ({ snuMail, googleAccessToken }) => {
-    const body = { snuMail, googleAccessToken };
-    const { status, data } = await apis['POST /user/signup/google']({ body });
-
-    if (status === 200) {
-      const accessToken = data.accessToken;
-
-      tokenLocalStorage.setToken({ token: accessToken });
-      tokenState.setToken({ token: accessToken });
-
-      return {
-        type: 'success',
-        data,
-      };
-    }
-    return { type: 'error', status, message: data.error };
-  },
-  googleSignIn: async ({ googleAccessToken }) => {
-    const body = { googleAccessToken };
-    const { status, data } = await apis['POST /user/signin/google']({ body });
-
-    if (status === 200) {
-      const accessToken = data.accessToken;
-
-      tokenLocalStorage.setToken({ token: accessToken });
-      tokenState.setToken({ token: accessToken });
 
       return {
         type: 'success',
@@ -172,7 +112,9 @@ export const implAuthService = ({
   },
   sendEmailCode: async ({ snuMail }) => {
     const body = { snuMail };
-    const { status, data } = await apis['POST /user/signup/send-code']({
+    const { status, data } = await apis[
+      'POST /user/snu-mail-verification/request'
+    ]({
       body,
     });
 
@@ -186,7 +128,9 @@ export const implAuthService = ({
   },
   verifyCode: async ({ code, snuMail }) => {
     const body = { snuMail, code };
-    const { status, data } = await apis['POST /user/signup/verify-email']({
+    const { status, data } = await apis[
+      'POST /user/snu-mail-verification/verify'
+    ]({
       body,
     });
 
@@ -199,8 +143,8 @@ export const implAuthService = ({
     return { type: 'error', status, message: data.error };
   },
   checkLocalIdDuplicate: async ({ localId }) => {
-    const body = { localId };
-    const { status, data } = await apis['POST /user/signup/id-duplicate']({
+    const body = { id: localId };
+    const { status, data } = await apis['POST /user/signup/check-id']({
       body,
     });
 
@@ -213,7 +157,7 @@ export const implAuthService = ({
     return { type: 'error', status, message: data.error };
   },
   reissueAccessToken: async () => {
-    const { status, data } = await apis['POST /user/token/refresh']();
+    const { status, data } = await apis['POST /user/refresh-token']();
 
     if (status === 200) {
       const accessToken = data.accessToken;
@@ -229,8 +173,10 @@ export const implAuthService = ({
     return { type: 'error', status, message: data.error };
   },
   checkGoogleEmail: async ({ token }) => {
-    const body = { googleAccessToken: token };
-    const { status, data } = await apis['POST /user/signup/google-email']({
+    const body = { accessToken: token };
+    const { status, data } = await apis[
+      'POST /user/snu-mail-verification/google-email'
+    ]({
       body,
     });
 
@@ -243,44 +189,10 @@ export const implAuthService = ({
     return { type: 'error', status, message: data.error };
   },
   logout: async ({ token }) => {
-    const { status, data } = await apis['POST /user/logout']({ token });
+    const { status, data } = await apis['POST /user/signout']({ token });
     tokenLocalStorage.removeToken();
     tokenState.removeToken();
     if (status === 200) {
-      return {
-        type: 'success',
-        data,
-      };
-    }
-    return { type: 'error', status, message: data.error };
-  },
-  addGoogleSignUp: async ({ snuMail, googleAccessToken }) => {
-    const body = { snuMail, googleAccessToken };
-    const { status, data } = await apis['POST /user/google']({ body });
-
-    if (status === 200) {
-      const accessToken = data.accessToken;
-
-      tokenLocalStorage.setToken({ token: accessToken });
-      tokenState.setToken({ token: accessToken });
-
-      return {
-        type: 'success',
-        data,
-      };
-    }
-    return { type: 'error', status, message: data.error };
-  },
-  addLocalSignUp: async ({ username, localId, password, snuMail }) => {
-    const body = { username, localId, password, snuMail };
-    const { status, data } = await apis['POST /user/local']({ body });
-
-    if (status === 200) {
-      const token = data.accessToken;
-
-      tokenLocalStorage.setToken({ token });
-      tokenState.setToken({ token });
-
       return {
         type: 'success',
         data,
