--- apps/client/src/shared/api/entities/index.ts
@@ -1,2 +1,0 @@
-export * from '@/shared/api/entities/params';
-export * from '@/shared/api/entities/schemas';
