--- apps/client/src/shared/api/index.ts
@@ -1,2 +1,0 @@
-export { implApi, type Apis } from '@/shared/api/client';
-export type { ExternalCallParams } from '@/shared/api/entities';
