+++ packages/api/src/entities/schemas.ts
@@ -5,6 +5,14 @@ type UserBriefDTO = {
   isAdmin: boolean;
 };
 
+type UserDTO = {
+  id: string;
+  snuMail: string;
+  username: string;
+  phoneNumber?: string;
+  isAdmin: boolean;
+};
+
 type AuthorBriefDTO = {
   id: string;
   name: string;
@@ -18,6 +26,7 @@ type RoleDTO = {
     | "FRONT"
     | "APP"
     | "BACKEND"
+    | "OTHERS"
     | "DESIGN"
     | "DATA"
     | "MARKETER";
@@ -25,22 +34,70 @@ type RoleDTO = {
   headcount: string;
 };
 
+type Link = {
+  link: string;
+  description: string;
+};
+
 type PostDTO = {
   id: string;
+  author: AuthorBriefDTO;
+
+  // 회사 정보
   companyName: string;
   email: string;
-  author: AuthorBriefDTO;
-  explanation: string;
-  tags: string[];
-  roles: RoleDTO[];
-  imageLink?: string;
+  slogan: string;
   investAmount?: number;
   investCompany: string[];
+  series: "SEED" | "PRE_A" | "A" | "B" | "C" | "D";
   IRDeckLink?: string;
   landingPageLink?: string;
-  externalDescriptionLink?: string[];
-  isActive: boolean;
+  imageLink?: string;
+  externalDescriptionLink?: Link[];
+  tags?: string[];
+
+  // post 정보
+  title: string;
+  isAlways: boolean;
   employmentEndDate: string;
+  createdAt: string;
+  updatedAt: string;
+  isActive: boolean;
+  category: RoleDTO["category"];
+  detail: string;
+  headcount: string;
+};
+
+export type PostBriefDTO = {
+  id: string;
+  author: AuthorBriefDTO;
+
+  // 회사 정보
+  companyName: string;
+  slogan: string;
+  investAmount?: number;
+  investCompany: string[];
+  series: "SEED" | "PRE_A" | "A" | "B" | "C" | "D";
+  imageLink: string;
+
+  // 포스트 정보
+  title: string;
+  isAlways: boolean;
+  employmentEndDate: string;
+  createdAt: string;
+  updatedAt: string;
+  isActive: boolean;
+  category: RoleDTO["category"];
+  headcount: string;
+};
+
+type ResumeDTO = {
+  id: string;
+  postId: string;
+  author: UserDTO;
+  content: string;
+  phoneNumber: string;
+  createdAt: string;
 };
 
 type LocalApplicantInfo = {
@@ -63,21 +120,6 @@ type SocialApplicantInfo = {
   token: string;
 };
 
-export type PostBriefDTO = {
-  id: string;
-  companyName: string;
-  email: string;
-  author: AuthorBriefDTO;
-  explanation: string;
-  tags: string[];
-  roles: RoleDTO[];
-  imageLink?: string;
-  investAmount?: number;
-  investCompany: string[];
-  isActive: boolean;
-  employmentEndDate: string;
-};
-
 // Params
 export type EchoParams = {
   message: string;
@@ -89,6 +131,10 @@ export type PostPathParams = {
 
 export type PostIdParams = {
   postId: string;
+};
+
+export type ResumeIdParams = {
+  resumeId: string;
 };
 
 // Request
@@ -133,6 +179,11 @@ export type IdRequest = {
 
 export type CreateAndUpdatePostRequest = PostDTO;
 
+export type ApplyCoffeeChatRequest = {
+  phoneNumber: string;
+  content: string;
+};
+
 // Response
 export type PretotypeUserSubmitResponse = {
   email: string;
@@ -169,3 +220,7 @@ export type TokenResponse = {
 export type GoogleEmailResponse = {
   googleEmail: string;
 };
+
+export type ResumeResponse = ResumeDTO;
+
+export type ResumeListResponse = { resumeList: ResumeDTO[] };
